requirements.txt
openpyxl==3.1.2
pandas==2.1.2
python-pptx==0.6.23
